requirements.txt
aerich==0.5.8
aioredis==2.0.0
aiosqlite==0.17.0
anyio==3.4.0
asgiref==3.4.1
async-timeout==4.0.1
asyncpg==0.25.0
click==8.0.3
databases==0.5.3
ddlparse==1.10.0
dictdiffer==0.9.0
fastapi==0.70.1
greenlet==1.1.2
h11==0.12.0
hiredis==2.0.0
idna==3.3
iso8601==0.1.16
pydantic==1.8.2
pypika-tortoise==0.1.1
pytz==2021.3
sniffio==1.2.0
SQLAlchemy==1.4.28
starlette==0.16.0
tortoise-orm==0.17.8
typing-extensions==4.0.1
uvicorn==0.16.0
passlib==1.7.4
bleach==4.0.0
fastapi_jwt_auth==0.5.0
fastapi-health==0.4.0
json-logging==1.3.0
fastapi-utils==0.2.1
httpx==0.19.0
asynch==0.1.9
pytest==7.1.1
requests==2.27.1
slowapi==0.1.5
Jinja2==3.0.1
pandas==1.4.2
email-validator==1.1.1
fastapi-mail==0.4.1
fastapi-admin==1.0.3
pandas-datareader==0.10.0
numpy==1.22.3
pyarrow==8.0.0
pandahouse==0.2.7
fastapi-pagination==0.9.3
pyparsing==3.0.9
fastapi-cprofile==0.0.2
memory-profiler==0.60.0
Pympler==1.0.1
scipy==1.8.1
pybind11==2.10.0
cppimport==22.8.2
Pillow==9.2.0